scripts/requirements.txt
mysql-connector-python==8.0.18
sendgrid==6.1.0
